silence gcc (pointer-to-int-cast)

+++ rewrite.c
@@ -806,7 +806,7 @@ void keep_field(spec)				   /*			     */
   Symbol* np;				   	   /*                        */
   Symbol field 	 = NO_SYMBOL;			   /*                        */
   Symbol pattern = NO_SYMBOL;		   	   /*                        */
-  int i;					   /*                        */
+  intptr_t i;					   /*                        */
  						   /*                        */
   sp_open(s);				   	   /*			     */
   if ((names = sp_symbols(&s)) == NULL)    	   /*		             */
@@ -846,7 +846,7 @@ void keep_field(spec)				   /*			     */
 			 field,			   /*                        */
 			 RULE_KEEP | RULE_REGEXP,  /*                        */
 			 true);	   	   	   /*                        */
-    i = (int)(*np) % K_RULES_SIZE;		   /*                        */
+    i = (intptr_t)(*np) % K_RULES_SIZE;		   /*                        */
     if (i < 0) i = -i;				   /*                        */
  						   /*                        */
     NextRule(rule) = k_rules[i];		   /*                        */
@@ -897,7 +897,7 @@ static bool dont_keep(sym,rec,db)		   /*                        */
   Record rec;					   /*                        */
   DB     db;					   /*                        */
 { Rule   r;					   /*                        */
-  int    idx = (int)(sym) % K_RULES_SIZE;	   /*                        */
+  intptr_t idx = (intptr_t)(sym) % K_RULES_SIZE;	   /*                        */
   if (idx < 0) idx = -idx;			   /*                        */
  						   /*                        */
   for (r = k_rules[idx]; r; r = NextRule(r))	   /*                        */
